+++ src/Components/PayPalExpress/PayPalExpress.php
@@ -7,11 +7,9 @@ namespace PaynlPayment\Shopware6\Components\PayPalExpress;
 use PayNL\Sdk\Model;
 use PayNL\Sdk\Exception\PayException;
 use PaynlPayment\Shopware6\Components\ExpressCheckoutUtil;
-use PaynlPayment\Shopware6\Enums\PaynlTransactionStatusesEnum;
 use PaynlPayment\Shopware6\Exceptions\PaynlPaymentException;
 use PaynlPayment\Shopware6\Exceptions\PayPalPaymentApi;
 use PaynlPayment\Shopware6\Repository\OrderDelivery\OrderDeliveryRepositoryInterface;
-use PaynlPayment\Shopware6\ValueObjects\PAY\Response\CreateOrderResponse;
 use PaynlPayment\Shopware6\ValueObjects\PayPal\Response\CreateOrderResponse as PayPalCreateOrderResponse;
 use PaynlPayment\Shopware6\ValueObjects\PayPal\Order\Amount as PayPalAmount;
 use PaynlPayment\Shopware6\ValueObjects\PayPal\Order\CreateOrder as PayPalCreateOrder;
@@ -219,20 +217,6 @@ class PayPalExpress
         return $orderResponse->getId();
     }
 
-    /** @throws Exception */
-    public function updatePaymentTransaction(CreateOrderResponse $orderResponse): void
-    {
-        $statusCode = $orderResponse->getStatus()->getCode();
-
-        $stateActionName = PaynlTransactionStatusesEnum::STATUSES_ARRAY[$statusCode] ?? null;
-
-        if (empty($stateActionName)) {
-            throw new Exception('State action name was not defined');
-        }
-
-        $this->processingHelper->instorePaymentUpdateState($orderResponse->getOrderId(), $stateActionName, $statusCode);
-    }
-
     /**
      * @throws PayPalPaymentApi
      * @throws PaynlPaymentException
@@ -306,11 +290,13 @@ class PayPalExpress
             $salesChannelContext->getContext()
         );
 
-        $this->updateOrder($orderTransaction->getOrder(), $payPalOrder, $salesChannelContext);
+        $order = $this->orderService->getOrderByNumber($orderTransaction->getOrder()->getOrderNumber(), $salesChannelContext->getContext());
 
-        $this->updateOrderCustomer($orderTransaction->getOrder()->getOrderCustomer(), $payPalOrder, $salesChannelContext);
+        $this->updateOrder($order, $payPalOrder, $salesChannelContext);
 
-        $this->updateCustomer($orderTransaction->getOrder()->getOrderCustomer()->getCustomer(), $payPalOrder, $salesChannelContext);
+        $this->updateOrderCustomer($order->getOrderCustomer(), $payPalOrder, $salesChannelContext);
+
+        $this->updateCustomer($order->getOrderCustomer()->getCustomer(), $payPalOrder, $salesChannelContext);
 
         $this->processingHelper->notifyActionUpdateTransactionByPayTransactionId($orderCreateResponse->getOrderId());